Fix for non-integer serials
When BIND has a problem loading a zone (e.g. due to an unreadable journal file)
the statistics page contains "-" for this zone as serial instead of an integer.
Up to now pybindxml fails with a ValueError in such case as it assumes that the
serial is an integer. This commit fixes that behavior by simply removing the
assumption that the serial has to be an integer.